pyproject.toml
[tool.poetry]
name = "rd-automation"
version = "0.1.0"
description = "RD Automation"
authors = ["Paulo Santiago"]

[tool.poetry.dependencies]
python = "^3.9"
python-decouple = "^3.8"
requests = "^2.27.1"
robotframework = "^7.1.1"
robotframework-appiumlibrary = "^2.1.0"
robotframework-faker = "^5.0.0"
robotframework-requests = "^0.9.2"
robotframework-seleniumlibrary = "^6.0.0"
robotframework-tidy = "^4.14.0"
webdriver-manager = "^4.0.2"

[tool.robocop]
exclude = ["0201","0202","0203","0503","0302"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"